Fix incorrect `lastPattern` value
In particular, mania-specific beatmaps that normally go via the
"passthrough" generator should not adjust the stored pattern value.

The "spinner" generator, which was previously intended to be used for
non-mania-specific beatmaps, is now valid even for mania-specific
beatmaps, and uses this value.

In other words, another way of writing this would be:
```csharp
if (conversion is SpinnerPatternGenerator || conversion is
PassThroughPatternGenerator) ? lastPattern : newPattern;
```